Improve Exception displayers and move the kernel classes to namespace \Foundation
