+++ system/Foundation/AliasLoader.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Core;
+namespace Foundation;
 
 class AliasLoader
 {
